[Refactoring ⚙️] 토글 훅 함수 로직, 타입가드방식 변경
